services/listener/requirements.txt
pika>=1.3.0
imapclient>=3.1.0
python-dotenv>=1.1.0,<2
